Wlan0 scan is denied by p2p busy traffic
:Release Notes:
Wlan0 scan is denied by p2p busy traffic

:Detailed Notes:
Wlan0 can't connect to AP when P2P is in Miracast.
It is because that wlan0 scan will be blocked due to P2P busy traffic.
With this patch, wlan0 scan will back to op channel after it scans 1 channel, and stay 100 ms in op channel.
It can make sure Miracast is stable and let wlan0 connect to AP during Miracast.

:Testing Performed:
[Legacy 1-113] is passed.

:Issues Addressed:
[RTKWBTV-171] 1st Veri TEST [Legacy 1-113] MCC WiFi connection

Change-Id: Ibcf031d4c8373c7d7ad8f4c884fd8ce07f2cd300