Enhance Layer 2 active roaming
:Release Notes:
Enhance Layer 2 active roaming

:Detailed Notes:
The default RTW_ROAM_RSSI_THRESHOLD is 70 set to 16
It means average signal is -85 dbm that will trigger roaming.

:Testing Performed:

:Issues Addressed:
Request form Pyeongtaek.
The condition is that there are 2 or 3 AP that have same SSID.
Case #1 : After connection with one AP, when the signal level is very lower, the device need to connect other AP.

Change-Id: I25c871b541f14c2f5948648ae5452d3baa4cdbd0